# helpers: Make UniqueStringsReuse allocation free